requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.26.0
peft>=0.8.0
bitsandbytes>=0.42.0
diffusers>=0.27.0
lpips>=0.1.4
torchmetrics>=1.3.0
gradio>=4.20.0
pandas>=2.0.0
pyyaml>=6.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
tqdm>=4.65.0
scipy>=1.11.0
jupyter
notebook
